pyproject.toml
[project]
name = "ptouch-bt"
version = "0.0.0"
description = "Use Brother P-touch label printers over Bluetooth on Linux"
requires-python = ">=3.10"
readme = "README.md"
authors = [{name = "Scott Hyndman", email = "scotty.hyndman@gmail.com"}]
dependencies = ["click>=8.1", "Pillow>=10", "pydantic>=2.13.3"]

[project.scripts]
ptouch = "ptouch_bt.cli.main:cli"

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"

[tool.basedpyright]
include = ["src"]
exclude = ["**/__pycache__"]
defineConstant = {DEBUG = true}
reportMissingImports = "error"
reportMissingTypeStubs = false

pythonVersion = "3.10"
pythonPlatform = "Linux"

executionEnvironments = [
  {root = "tests", reportPrivateUsage = false},
  {root = "src"},
]

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

line-length = 88
indent-width = 2
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
